Part: Move OCC conversions into their own header
Tools.h mixed the Base vec_traits specialisations for gp_Pnt, gp_Vec, gp_Dir
and gp_XYZ with the Part free functions and the Tools class. Consumers that
wanted only a conversion pulled in TopoShape, Bnd_Box, Geom_Surface and a
dozen more OCC headers; consumers that wanted only Tools:: parsed four trait
specialisations they never used.

The traits move to Part/App/Conversion/OCC.h, matching Gui/Conversion, and
each consumer includes the one it needs. GizmoHelper.cpp had been reaching
TopoDS_Edge through Tools.h and now includes it directly.
