+++ src/Mod/Measure/App/MeasureAngle.cpp
@@ -47,6 +47,7 @@
 #include "MeasureAngle.h"
 #include <Mod/Part/App/Geometry.h>
 #include <Mod/Part/App/PartFeature.h>
+#include <Mod/Part/App/Conversion/OCC.h>
 #include <Mod/Part/App/Tools.h>
 #include <Mod/Part/App/Datums.h>
 #include <TopTools_IndexedMapOfShape.hxx>
