+++ src/Mod/Part/App/ChFi2d/ChFi2d_FilletAlgoPyImp.cpp
@@ -37,7 +37,7 @@
 #include "TopoShapeEdgePy.h"
 #include "TopoShapeWirePy.h"
 #include "PlanePy.h"
-#include "Tools.h"
+#include "Conversion/OCC.h"
 
 
 using namespace Part;
