+++ src/Mod/Part/App/GizmoHelper.cpp
@@ -41,7 +41,8 @@
 
 #include <Base/BoundBox.h>
 #include <Base/Converter.h>
-#include <Mod/Part/App/Tools.h>
+#include <Mod/Part/App/Conversion/OCC.h>
+#include <TopoDS_Edge.hxx>
 
 
 EdgeMidPointProps getEdgeMidPointProps(Part::TopoShape& edge)
