+++ src/Mod/Part/App/LinearPatternExtension.cpp
@@ -34,7 +34,7 @@
 #include <TopoDS_Face.hxx>
 #include <App/Datums.h>
 #include <Base/Axis.h>
-#include <Mod/Part/App/Tools.h>
+#include <Mod/Part/App/Conversion/OCC.h>
 #include <Mod/Part/App/TopoShape.h>
 #include <Mod/Part/App/Part2DObject.h>
 #include <App/DocumentObject.h>
