+++ src/Mod/Part/App/PolarPatternExtension.cpp
@@ -35,7 +35,6 @@
 #include <App/Datums.h>
 #include <Base/Axis.h>
 #include <Base/Tools.h>
-#include <Mod/Part/App/Tools.h>
 #include <Mod/Part/App/TopoShape.h>
 #include <Mod/Part/App/Part2DObject.h>
 #include <App/DocumentObject.h>
