+++ src/Mod/Part/Gui/DlgPrimitives.cpp
@@ -52,7 +52,7 @@
 #include <Mod/Part/App/PrimitiveFeature.h>
 #include <Mod/Part/App/FeaturePartBox.h>
 #include <Mod/Part/App/FeaturePartCircle.h>
-#include <Mod/Part/App/Tools.h>
+#include <Mod/Part/App/Conversion/OCC.h>
 
 #include "DlgPrimitives.h"
 #include "ui_DlgPrimitives.h"
