+++ src/Mod/Part/Gui/ViewProviderPreviewExtension.cpp
@@ -38,7 +38,6 @@
 #include <Gui/Conversion/Coin.h>
 #include <Gui/Inventor/So3DAnnotation.h>
 #include <Mod/Part/App/PreviewExtension.h>
-#include <Mod/Part/App/Tools.h>
 #include <Mod/Part/App/TopoShapePy.h>
 
 using namespace PartGui;
