+++ src/Mod/PartDesign/Gui/ViewProviderHole.cpp
@@ -48,6 +48,7 @@
 #include <App/Material.h>
 #include <Gui/Application.h>
 #include <Gui/ViewProvider.h>
+#include <Mod/Part/App/Conversion/OCC.h>
 #include <Mod/Part/App/Tools.h>
 #include <Mod/PartDesign/App/Body.h>
 #include <Mod/PartDesign/App/Feature.h>
